rename Effect::child() to Effect::collection()
