+++ properties/EffectChildAddOnAllAggregates.php
@@ -65,7 +65,7 @@ final class EffectChildAddOnAllAggregates implements Property
                 $manager
                     ->repository(User::class)
                     ->effect(
-                        Effect::child('addresses')->add(
+                        Effect::collection('addresses')->add(
                             User\Address::new($this->address),
                         ),
                     ),
